Rename gmfunction pkg to api pkg
