+++ api/db.go
@@ -1,4 +1,4 @@
-package gmfunction
+package api
 
 import (
 	"io/ioutil"
