+++ api/get_user.go
@@ -1,4 +1,4 @@
-package gmfunction
+package api
 
 import (
 	"io/ioutil"
