+++ go.mod
@@ -5,7 +5,7 @@ go 1.17
 require (
 	github.com/go-sql-driver/mysql v1.6.0
 	github.com/gorilla/mux v1.8.0
-	local.packages/gmfunction v0.0.0-00010101000000-000000000000
+	local.packages/api v0.0.0-00010101000000-000000000000
 )
 
 require (
@@ -38,4 +38,4 @@ replace local.packages/gmtoken => ./gmtoken
 
 replace local.packages/transaction => ./transaction
 
-replace local.packages/gmfunction => ./gmfunction
+replace local.packages/api => ./api
