+++ main.go
@@ -9,7 +9,7 @@ import (
 	"math/rand"
 	"net/http"
 	"github.com/gorilla/mux"
-	gmfunction "local.packages/gmfunction"
+	api "local.packages/api"
 	_ "github.com/go-sql-driver/mysql"
 )
 
@@ -19,13 +19,13 @@ func main() {
 	// URLと処理
 	router.HandleFunc("/", home)
 	// ユーザ関連API
-	router.HandleFunc("/user/create", gmfunction.CreateUser).Methods("POST")
-	router.HandleFunc("/user/get", gmfunction.GetUser).Methods("GET")
-	router.HandleFunc("/user/update", gmfunction.UpdateUser).Methods("PUT")
+	router.HandleFunc("/user/create", api.CreateUser).Methods("POST")
+	router.HandleFunc("/user/get", api.GetUser).Methods("GET")
+	router.HandleFunc("/user/update", api.UpdateUser).Methods("PUT")
 	// ガチャ関連API
-	router.HandleFunc("/gacha/draw", gmfunction.DrawGacha).Methods("POST")
+	router.HandleFunc("/gacha/draw", api.DrawGacha).Methods("POST")
 	// キャラクター関連API
-	router.HandleFunc("/character/list", gmfunction.GetCharacterList).Methods("GET")
+	router.HandleFunc("/character/list", api.GetCharacterList).Methods("GET")
 	// 乱数のシード値を設定
 	seed, _ := crand.Int(crand.Reader, big.NewInt(math.MaxInt64))
 	rand.Seed(seed.Int64())
